refactor: extract shared generateId and resource ID validation into lib/id-helpers.js
- Create backend/lib/id-helpers.js with generateId(), isValidResourceId(),
  validateIdParam(), and RESOURCE_ID_RE
- Remove duplicate generateId() from alerts.js, webhooks.js, annotations.js
- Remove duplicate RESOURCE_ID_RE / SAFE_ID_RE / WEBHOOK_ID_RE regexes
- Remove redundant per-route validateIdParam() middleware in alerts.js
  (router.param already handles the same validation)
- Annotations uses generateId('ann') prefix via wrapper function
